+++ libavcodec/x86/dsputil_avg_template.c
@@ -27,14 +27,14 @@
 //FIXME the following could be optimized too ...
 static void DEF(ff_put_no_rnd_pixels16_x2)(uint8_t *block,
                                            const uint8_t *pixels,
-                                           int line_size, int h)
+                                           ptrdiff_t line_size, int h)
 {
     DEF(ff_put_no_rnd_pixels8_x2)(block,     pixels,     line_size, h);
     DEF(ff_put_no_rnd_pixels8_x2)(block + 8, pixels + 8, line_size, h);
 }
 
 static void DEF(ff_put_pixels16_y2)(uint8_t *block, const uint8_t *pixels,
-                                    int line_size, int h)
+                                    ptrdiff_t line_size, int h)
 {
     DEF(ff_put_pixels8_y2)(block,     pixels,     line_size, h);
     DEF(ff_put_pixels8_y2)(block + 8, pixels + 8, line_size, h);
@@ -42,35 +42,35 @@ static void DEF(ff_put_pixels16_y2)(uint8_t *block, const uint8_t *pixels,
 
 static void DEF(ff_put_no_rnd_pixels16_y2)(uint8_t *block,
                                            const uint8_t *pixels,
-                                           int line_size, int h)
+                                           ptrdiff_t line_size, int h)
 {
     DEF(ff_put_no_rnd_pixels8_y2)(block,     pixels,     line_size, h);
     DEF(ff_put_no_rnd_pixels8_y2)(block + 8, pixels + 8, line_size, h);
 }
 
 static void DEF(ff_avg_pixels16)(uint8_t *block, const uint8_t *pixels,
-                                 int line_size, int h)
+                                 ptrdiff_t line_size, int h)
 {
     DEF(ff_avg_pixels8)(block,     pixels,     line_size, h);
     DEF(ff_avg_pixels8)(block + 8, pixels + 8, line_size, h);
 }
 
 static void DEF(ff_avg_pixels16_x2)(uint8_t *block, const uint8_t *pixels,
-                                    int line_size, int h)
+                                    ptrdiff_t line_size, int h)
 {
     DEF(ff_avg_pixels8_x2)(block,     pixels,     line_size, h);
     DEF(ff_avg_pixels8_x2)(block + 8, pixels + 8, line_size, h);
 }
 
 static void DEF(ff_avg_pixels16_y2)(uint8_t *block, const uint8_t *pixels,
-                                    int line_size, int h)
+                                    ptrdiff_t line_size, int h)
 {
     DEF(ff_avg_pixels8_y2)(block,     pixels,     line_size, h);
     DEF(ff_avg_pixels8_y2)(block + 8, pixels + 8, line_size, h);
 }
 
 static void DEF(ff_avg_pixels16_xy2)(uint8_t *block, const uint8_t *pixels,
-                                     int line_size, int h)
+                                     ptrdiff_t line_size, int h)
 {
     DEF(ff_avg_pixels8_xy2)(block,     pixels,     line_size, h);
     DEF(ff_avg_pixels8_xy2)(block + 8, pixels + 8, line_size, h);
